+++ src/vmm/src/vmm_config/machine_config.rs
@@ -55,21 +55,20 @@ impl fmt::Display for VmConfigError {
 #[serde(deny_unknown_fields)]
 pub struct VmConfig {
     /// Number of vcpu to start.
+    #[serde(deserialize_with = "deserialize_vcpu_num")]
+    pub vcpu_count: u8,
+    /// The memory size in MiB.
+    pub mem_size_mib: usize,
+    /// Enables or disabled SMT.
+    #[serde(default, deserialize_with = "deserialize_smt")]
+    pub smt: bool,
+    /// A CPU template that it is used to filter the CPU features exposed to the guest.
     #[serde(
         default,
-        skip_serializing_if = "Option::is_none",
-        deserialize_with = "validate_vcpu_num"
+        deserialize_with = "deserialize_cpu_template",
+        skip_serializing_if = "CpuFeaturesTemplate::is_none"
     )]
-    pub vcpu_count: Option<u8>,
-    /// The memory size in MiB.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mem_size_mib: Option<usize>,
-    /// Enables or disabled SMT.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub smt: Option<bool>,
-    /// A CPU template that it is used to filter the CPU features exposed to the guest.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cpu_template: Option<CpuFeaturesTemplate>,
+    pub cpu_template: CpuFeaturesTemplate,
     /// Enables or disables dirty page tracking. Enabling allows incremental snapshots.
     #[serde(default)]
     pub track_dirty_pages: bool,
@@ -78,10 +77,10 @@ pub struct VmConfig {
 impl Default for VmConfig {
     fn default() -> Self {
         VmConfig {
-            vcpu_count: Some(1),
-            mem_size_mib: Some(DEFAULT_MEM_SIZE_MIB),
-            smt: Some(false),
-            cpu_template: None,
+            vcpu_count: 1,
+            mem_size_mib: DEFAULT_MEM_SIZE_MIB,
+            smt: false,
+            cpu_template: CpuFeaturesTemplate::None,
             track_dirty_pages: false,
         }
     }
@@ -89,35 +88,149 @@ impl Default for VmConfig {
 
 impl fmt::Display for VmConfig {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let vcpu_count = self.vcpu_count.unwrap_or(1);
-        let mem_size = self.mem_size_mib.unwrap_or(DEFAULT_MEM_SIZE_MIB);
-        let smt = self.smt.unwrap_or(false);
-        let cpu_template = self
-            .cpu_template
-            .map_or("Uninitialized".to_string(), |c| c.to_string());
         write!(
             f,
             "{{ \"vcpu_count\": {:?}, \"mem_size_mib\": {:?}, \"smt\": {:?}, \
              \"cpu_template\": {:?}, \"track_dirty_pages\": {:?} }}",
-            vcpu_count, mem_size, smt, cpu_template, self.track_dirty_pages
+            self.vcpu_count, self.mem_size_mib, self.smt, self.cpu_template, self.track_dirty_pages
         )
     }
 }
 
-fn validate_vcpu_num<'de, D>(d: D) -> std::result::Result<Option<u8>, D::Error>
-where
-    D: de::Deserializer<'de>,
-{
-    let val = Option::<u8>::deserialize(d)?;
-    if let Some(ref value) = val {
-        if *value > MAX_SUPPORTED_VCPUS {
-            return Err(de::Error::invalid_value(
-                de::Unexpected::Unsigned(u64::from(*value)),
-                &"number of vCPUs exceeds the maximum limitation",
-            ));
+/// Spec for partial configuration of the machine.
+/// This struct mirrors all the fields in `VmConfig`.
+/// All fields are optional, but at least one needs to be specified.
+/// If a field is `Some(value)` then we assume an update is requested
+/// for that field.
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(deny_unknown_fields)]
+pub struct VmUpdateConfig {
+    /// Number of vcpu to start.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        deserialize_with = "deserialize_vcpu_num"
+    )]
+    pub vcpu_count: Option<u8>,
+    /// The memory size in MiB.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub mem_size_mib: Option<usize>,
+    /// Enables or disabled SMT.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        deserialize_with = "deserialize_smt"
+    )]
+    pub smt: Option<bool>,
+    /// A CPU template that it is used to filter the CPU features exposed to the guest.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        deserialize_with = "deserialize_cpu_template"
+    )]
+    pub cpu_template: Option<CpuFeaturesTemplate>,
+    /// Enables or disables dirty page tracking. Enabling allows incremental snapshots.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub track_dirty_pages: Option<bool>,
+}
+
+impl VmUpdateConfig {
+    /// Checks if the update request contains any data.
+    /// Returns `true` if all fields are set to `None` which means that there is nothing
+    /// to be updated.
+    pub fn is_empty(&self) -> bool {
+        if self.vcpu_count.is_none()
+            && self.mem_size_mib.is_none()
+            && self.cpu_template.is_none()
+            && self.smt.is_none()
+            && self.track_dirty_pages.is_none()
+        {
+            return true;
+        }
+
+        false
+    }
+}
+
+impl From<VmConfig> for VmUpdateConfig {
+    fn from(cfg: VmConfig) -> Self {
+        VmUpdateConfig {
+            vcpu_count: Some(cfg.vcpu_count),
+            mem_size_mib: Some(cfg.mem_size_mib),
+            smt: Some(cfg.smt),
+            cpu_template: Some(cfg.cpu_template),
+            track_dirty_pages: Some(cfg.track_dirty_pages),
         }
     }
+}
+
+/// Deserialization function for the `vcpu_num` field in `VmConfig` and `VmUpdateConfig`.
+/// This is called only when `vcpu_num` is present in the JSON configuration.
+/// `T` can be either `u8` or `Option<u8>` which both support ordering if `vcpu_num` is
+/// present in the JSON.
+fn deserialize_vcpu_num<'de, D, T>(d: D) -> std::result::Result<T, D::Error>
+where
+    D: de::Deserializer<'de>,
+    T: Deserialize<'de> + PartialOrd + From<u8>,
+{
+    let val = T::deserialize(d)?;
+
+    if val > T::from(MAX_SUPPORTED_VCPUS) {
+        return Err(de::Error::invalid_value(
+            de::Unexpected::Other(&"vcpu_num"),
+            &"number of vCPUs exceeds the maximum limitation",
+        ));
+    }
+    if val < T::from(1) {
+        return Err(de::Error::invalid_value(
+            de::Unexpected::Other(&"vcpu_num"),
+            &"number of vCPUs should be larger than 0",
+        ));
+    }
+
     Ok(val)
+}
+
+/// Deserialization function for the `smt` field in `VmConfig` and `VmUpdateConfig`.
+/// This is called only when `smt` is present in the JSON configuration.
+fn deserialize_smt<'de, D, T>(d: D) -> std::result::Result<T, D::Error>
+where
+    D: de::Deserializer<'de>,
+    T: Deserialize<'de> + PartialEq + From<bool>,
+{
+    let val = T::deserialize(d)?;
+
+    // If this function was called it means that `smt` was specified in
+    // the JSON. On aarch64 the only accepted value is `false` so throw an
+    // error if `true` was specified.
+    #[cfg(target_arch = "aarch64")]
+    if val == T::from(true) {
+        return Err(de::Error::invalid_value(
+            de::Unexpected::Other(&"smt"),
+            &"Enabling simultaneous multithreading is not supported on aarch64",
+        ));
+    }
+
+    Ok(val)
+}
+
+/// Deserialization function for the `cpu_template` field in `VmConfig` and `VmUpdateConfig`.
+/// This is called only when `cpu_template` is present in the JSON configuration.
+fn deserialize_cpu_template<'de, D, T>(_d: D) -> std::result::Result<T, D::Error>
+where
+    D: de::Deserializer<'de>,
+    T: Deserialize<'de>,
+{
+    // If this function was called it means that `cpu_template` was specified in
+    // the JSON. Return an error since `cpu_template` is not supported on aarch64.
+    #[cfg(target_arch = "aarch64")]
+    return Err(de::Error::invalid_value(
+        de::Unexpected::Enum,
+        &"CPU templates are not supported on aarch64",
+    ));
+
+    #[cfg(target_arch = "x86_64")]
+    T::deserialize(_d)
 }
 
 /// Template types available for configuring the CPU features that map
@@ -128,6 +241,15 @@ pub enum CpuFeaturesTemplate {
     C3,
     /// T2 Template.
     T2,
+    /// No CPU template is used.
+    None,
+}
+
+/// Utility methods for handling CPU template types
+impl CpuFeaturesTemplate {
+    fn is_none(&self) -> bool {
+        *self == CpuFeaturesTemplate::None
+    }
 }
 
 impl fmt::Display for CpuFeaturesTemplate {
@@ -135,7 +257,14 @@ impl fmt::Display for CpuFeaturesTemplate {
         match self {
             CpuFeaturesTemplate::C3 => write!(f, "C3"),
             CpuFeaturesTemplate::T2 => write!(f, "T2"),
+            CpuFeaturesTemplate::None => write!(f, "None"),
         }
+    }
+}
+
+impl Default for CpuFeaturesTemplate {
+    fn default() -> Self {
+        CpuFeaturesTemplate::None
     }
 }